groups: restore and simplify Join behavior
This change removes extraneous codepaths from the Group Join flow, and also restores behavior as specified in the Figma designs - when clicking the Join button from the Find Groups search results, it opens the JoinGroupModal for a preview of the group instead of directly navigating.

This might also resolve #486, will confirm with further testing.

TODO: consume the backend endpoint to reject a group invite