heaps: HeapInput component
This change introduces a new HeapInput component to allow a user to submit a Link or Text curio to a heap (aka a Collection). It also updates several tangential items: updating addCurio and delCurio to be async, using the ChannelIcon component in ChannelHeader, introducing a new hook (useNest), and refactoring the existing in-flight add curio flow.