heaps: edit curio modal and flow
This change allows a User to edit a curio from the HeapChannel view (when viewing the Grid or List view).

It updates types, adds new components, and introduces a new hook.

This is most of #567. When testing locally, the backend API call is failing, so it's not ready for review / merge just yet.