heaps: edit curio modal uses Rich Text
This change allows Curio edits to now be made with Rich Text.

There is one oustanding issue with the tiptap#parseInline function. For some Curio content, it is not currently possible to deserialize the persisted value back into JSON. So when clicking Edit on one of these, the modal will show a blank text field.