ui: update comment types to latest, and rework staggered fetching
We keep a map of noteId to either a note or an outline. An outline is a
truncated version of a note which has just enough to render the fields
for the list view. When we click into a note, we display a loading
indicator while we load the full note. Also moves feature set inline
with chat messages