groups: 0-to-1 migration [wip]
This uses the migration logic innovated by @liam-fitzgerald in `diary.hoon`. It migrates a groups 0 state to groups 1 state (with secret boolean).

This is a WIP commit: it "almost" works; currently crashingon line 65 due to a `nest-fail`:
```
/app/groups/hoon::[65 17].[65 43]>
```

Paired with Hunter on this yesterday, but now pivoted to another high priority Bridge issue. Will either hand this off or return to tihs after resolving Bridge.