channels: fix recent sort in sidebar
This change adds an effect hook to prefetch messages. When selecting a group, the messages for each channel will be fetched, then the last message timestamp will be compared to determine the sort order.

TODO: there is likely a more efficient way to do this, going to do some testing to confirm