scroller: remount when switching channels
This change forces a scroller remount for each channel switch.
This resets the scroll position when switching channels;
previously, when switching between channels, the virtuoso
internal scroll index would remain the same. So, if one scrolled
far back in a long channel, then switched to a less active one,
the channel would be scrolled to the top.