perf: dynamic scroller page fetching
Since we pre-fetch all messages from all channels, this change decreases the initial page fetch size to reduce load on the backend, and to render the page sooner.

Then for future page fetches, it uses a much higher number, to reduce subsequent fetches as the user backscrolls.

In a future commit, we will hook into the virtuoso scroll seek configuration; if the user is scrolling above a certain threshold velocity, we will fetch even more messages in anticipation of them going deep in the backscroll.