migration: %bait placeholder
Adds a `%bait` field to `%cite` representing an imported reference.

a `%bait` will only appear through the migrate flow and only when the
group being imported is not the same as the group holding the referenced
source. This disctinction must be made because `%cite` may not contain
enough information to reconstruct the imported reference, since the host
of the given reference may not have migrated yet.

Thus, `%bait` is primarily a placeholder for a potentially unresolvable
reference that will be handled at a later date by the `%nark` app. This
will allow us to begin the migration without losing the necessary data
needed to reconstruct refs.

(This will be available to anyone who migrates, so an enterprising
developer could theoretically create custom reference types for
imported instances.)