ux: support color hexes starting with 0
This resolves #1390 by changing how `@ux` values are normalized in the
frontend. When a color hex such as `#00FFFF` is persisted in the
backend, it is converted to a `@ux` (unsigned hexadecimal). The
extraneous 0 values are not saved, which then requires them to be
restored when loaded in the frontend.