talk-cli: scry for dms and clubs
We've added two arms (+get-dms and +get-clubs) which scry for accepted and pending dm identifiers, and all club identifiers. We take in pending dms because we plan on building functionality to interact with dm invites. The endpoint for clubs is total so we just grab all of it, regardless of status. $crew information come with the scry, which we can use later.

We've kept these separate because while dms and chats come in the form of sets, clubs show up in a map.

+target-exists has been updated to handle any $whom:chat.

Additionally, we've added +dms and +clubs, to accompany +chats, which call the set of +get arms so they can display their targets when the commands, ;dms or ;clubs, are passed.