talk-cli: subscribes to dms and clubs
The `+connect` arm has been renamed to distinguish between `+dm-connect` and `+club-connect`, which build lists of `%watch` cards for their respective subscriptions. The reason we separated these was because dms and clubs require specific ship (`/dm/<ship>/ui`) and club-id (`/club/<club-id>/ui`)  subscriptions. Opposed to chats, which only require a single subscription (`/chat/ui`) to the backend to receive message updates across chats.

So in `+dm-connect` and `+club-connect`, we check our available dms and chats against `wex.bowl` to determine if we're already subscribed. If not, we'll add a `%watch` card in our list to do so.

These connect arms are called in `+poke-noun`, `+prep:tc`  to establish or reestablish subscriptions if we need to, and of course in `+on-agent` when we're kicked from any wires.