talk-cli: handles $dm and $club %facts
New dm and club messages come in the form of `%writ-diff` marks, on `/dm/<ship>/ui` and `/club/<club-id>/ui` wires respectively. These are passed to `+on-dm-update` and `+on-club-update` arms to process message additions. Right now, we only care about adding messages so within each arm we check the `$diff:writs` for `%add` and output the message to session, if true.

To avoid copypasta, we broke out the portion of `+on-chat-update` that updated the session (now `+update-session`) so that all three update arms could use it.