talk-cli: portion of +read-post now checks history for dupes
Sending a `%club-action` poke to `%chat` requires that we define the `echo:club` number. Because `echo:club` is incremented upon each pass to `+gossip:cu-pass:cu` until it reaches 2, I assumed this pass should start with 0.

However, this presented a problem because this meant that we now received 2 `%facts` of the same message after sending the poke, and both of which were being output to session. The Talk frontend must have some mechanism to prevent this so I did the same for %talk-cli. `+read-post` now checks the `history` and discards any duplicate `[whom:chat id:chat]` identifiers.

Right now, it checks the whole history, which is inefficent so I'm thinking it might be wiser to just check the previous x identifiers since it may be safe to assume any duplicate facts will arrive within a short time window (?).