talk-cli: +parser takes dm and club messages
We specifically wanted to be able to send a dm with `;<ship> <message>` so the first thing the parser attempts after matching a mic is exactly this. Because the `%say` command previously only required a `(list inline:chat)`, and we now need to distinguish between a `%say` command that is directed to a channel we're `viewing` and a dm, outside of the current context, we included a `(unit ship)` within the command.

So the parser then attempts to match a ship and message (found in the `+dm:parser` arm) and builds a `%say` command with the ship if it does. If it's just a `;<ship>` without a message, `+targ` will take over indicating an `audience` change. If there's no `;<ship>` before the message, the text will be passed to `+message` as before, and an empty unit will be put into the `%say` command indicating `+say` should post the message in the current `audience`.

This of course required `$whom` support, so I updated the `+targ` arm to include a `+whom` parser, which uses existing parsers such as `+ship` and `+name`, and a new one (`+club-id`).

I commented out a few non-essential arms/gates which I'll return to fix:
- `+tarl` was commented out since it currently consumes only a `$flag:chat`.
-  `+ships` and `+tars` because I haven't adjust them to support `$whom` yet, and they've aren't currently called anywhere.
- the url parser in `+content` because I was getting an error