talk-cli: +say sends dm and club messages
The crux of this change comes down to `+say` now needing to distinguish between a message directed at a channel we're currently viewing (i.e. just a message post), and one directed at a specific ship outside of the current context, with the command `;<ship> <message>`.

As mentioned in the previous commit, a `(unit ship)`  was added to the `%say` command so we simply branch on whether this is empty or not. When it's empy, `+say` just checks the current audience, it's `$whom` type, and then passes the message to `%chat`. If the unit is full, it changes the audience (currently only to a `target` that exists), adds it to `viewing`, binds a `glyph` (if needed), changes the prompt, and sends the message.

We wanted to do this in one fell swoop becuase it seems natural that if someone is passing `;<ship> <message>` they're wanting to fully change contexts.

I currently have an `rsvp` implementation commented out due to hitting an error and running out of time, but soon the `+say` arm will be able to pass an `rsvp` response for `pending-dms`, as well as, being able to send a dm without needing the target to exist beforehand.