talk-cli: +dm-connect and +club-connect arms are called in +view, +dms, and +clubs
So that we're sure we are subscribed to any dms or club threads that have come in since loading the agent.

I figured calling one of these three arms, which reflect the commands, `;view <target>`, `;dms`, and `;clubs`, are key moments when we may have a new dm or club thread that needs subscribed to. Not included in this commit is ajustmenst to `+say` to perform the same thing, particularly when we establish a new dm with `;<ship> <message>` (implemenation notes to come in the next commit).