talk-cli: establish a dm and send an $rsvp response with `;<ship> <message>`
We couldn't previously send an unsolicited dm with `;<ship> <message>` or do anything to a dm we didn't initialize. Now `+say` takes care of accepting a pending dm and establishing dm. If `+say` has a `ship` passed through it, it no longer checks if the `+target-exists` before adding to `audience`, `viewing`, bindng a glyph, and sending the message.

Additionally, I added a possible `%watch` card to `+say` instead of calling `+dm-connect`. I didn't use it because `+dm-connect` calls `+get-accepted-dms` to determine whether it should pass a watch card or not. If we're establishing a new message with a ship, the new dm doesn't reach the `accepted-dms` state in `%chat` fast enough for `+dm-connect` to realize it needs to send a `%watch` card. So I opted for including a `%watch` card in our list of `+say` cards, if we're not previously subscribed.

The possible `$rsvp` card is simply a `%dm-rsvp` poke, and we check our `pending-dms` to determine whether to include or not.