talk-cli: archived clubs are removed the +get-clubs map
We did not distinguish between club statuses before so when we passed `;clubs` all existing ones, including archived, appeared in the list. Ideally, we just see the active and pending ones, so I added a filter.

By doing this, I discovered a bug in `%chat` where essentially the `net:club` status doesn't change upon our being invited to a new chat, or leaving one. I'll add it as an issue to the main repo.

Currently, the filter checks the `team:club` and `hive:club` and if both are null we know it's archived.