talk-cli: club members are displayed next to id
When rendering a list of club ids, we append club members for ease of reference. I added a couple sub-arms to `+full:tr` so that existing arms that call `phat:tr`, which defers to `+full:tr` when a club is passed, will append the ships. `+show-targets` (called by `+view`,  `+targets`, `+clubs`) now defers to `+full:tr` specifically for clubs.

Currently, there's too much space rendered between each ship. I'll fix this as we get into the rendering portion of the apprenticeship.