talk-cli: added +meta arm in +tr
I originally had the meta data rendering functions for club ids in the `+full:tr` arm, but figured it's wiser to have this separated into a general `+meta` arm for clarity, and more general usage. The other reason I broke this out was so other arms such as `+show:tr` and `phat:tr` are free to simply render targets - I was previously getting a crash when `+prompt` was being called as it did not like the additional data appended to the target.

Arms such as `+show-target` and `+lookup-glyph` now call `+meta:tr` instead of `+phat:tr` so that meta data for club ids are displayed.

Chats and dms could also use it down the road, if there's a need for it later.

Within `+meta:tr` we branch on a type union where `%club` will render the club id and either the club name (if there is one) or the first 4 ships with a member count if there's more.