talk-cli: subscriptions only for targets in view
Initially subscriptions for dms and clubs were established for all existing targets upon loading the agent, but for efficency purposes, we've changed the model to only maintain a subscription while a given target is in view.

This includes chats, which were previously all established and maintained on a mono-subscription wire, `/chat/ui`. Now they work like dms and clubs, and are maintained on the target level.

Overall, this update entailed consolidating the channel-based connect arms back into a single `+connect` arm, which just branches on the target type.

This also meant updating  `+flee` to unsubscribe from targets we're no longer `viewing`.