talk-cli: +make-bind, +subscribe, +compose-channel, and +switch-channel
I dispersed functions that were primarily taking place in `+view` and `+say` because the amount of card creation and state manipulation was becoming unweildy. Now, by the end of `+view` and `+say`, there's simply a call to `+switch-channel`, which utillizes `+make-bind` (glyph binding), `+subscribe` (send watch card), and `+compose-channel` (change audience, put in view, and build message history).

I'm working on a fix for the crash that occurs when passing `;<ship> <message>` on a ship that isn't already in our dms. Basically, the scry crashes because the poke happens after. Conversely, if we have an existing dm with a ship and we send a message this way, the message is added to history in the correct order, but it's rendered at the top of the 20 messages pulled in. In this case I need to delay the poke, and have the scry go first.