talk-cli: render latest 20 messages
Now, by putting a target in view, we pull in the latest 20 messages so we can see where the conversation left off. `+view` and `+say` call the `+build-history` arm to handle scrying for the messages (`+get-messages`) and passing them through `+read-post` to output them to session in order.

The challenge implementing this new functionality was inserting the reference number for scried messages already in `history`. Because *new*-message reference numbers are pulled from `count`, I made a separate method to grab reference numbers already in `history`. First, I added a `+number-reference` arm to find the index for a given pointer, which doubles as a way for `+read-post` to know whether the message being passed through is new - and thus added to history - or already held and just output to session. `+read-post` still relies on `+show-post`, which has been modified to accept an `index=(unit @ud)` and `print-index=?`, which are used to know whether it should print an index number given to it, or the current `count` number.

Because `+build-history` has the list of messages it's passing through `+read-post`/`+show-post`, I thought it makes sense for it to also keep track of when to print the reference number. However, I'm not entirely satisfied with this implementation as it feels like this should be `+show-post`'s job. I mostly did it this way because `+show-post` only accepts a single message and I needed a way for it to output existing indexes without solely relying on `count` - perhaps, it might be worth expanding `+show-post` (or `+read-post`) to accept a list of messages now that we're not just outputting `%fact`s upon receival?