talk-cli: messages aren't special
Initially, I added some functionality to check whether a message existed in `history` or not, but after meeting with Mark, we concluded it was best to keep it simple and just add messages to the `history` list. This solves the sequential pointer reference issue which made it basically impossible to know some message's pointer numbers from the rendered thread.

From a performance perspective, the issue with this implementation is that the `history` list will balloon pretty quickly as messages are added via scry each time we switch channels. From a user's pov, the pointer number for a given message will change each time they hop to and from a channel. Perhaps, I should implement some kind of duplicate trimmer that runs occasionally?