talk-cli: dm poke and scry resolution speed
Because the command `;<ship> <message>` causes a scry to occur (in the case of changing the channel) and sends a poke, we need to delay one before the other. If the dm is established we scry first then send the poke so our new message displays at the bottom of the thread. For unestalbished dms, we poke first then scry so that our scry doesn't fail.

The former case is straight forward as we simply make sure our poke card resolves first. In the latter case, we set a behn timer to delay the scry and ensure we get our message.