talk-cli: message existence check
To go with the `%u` scry in `%chat`, I added an arm in `%talk-cli` to perform it. I followed the pattern of dropping a sub-arm for scries at the bottom, which specifically accepts `%u`s.

Also, I couldn't break the `+forge` part up into a separate commit so it's included. I just drop it from `+select` so other arms could use it.