talk-cli: reworked +select and added +pointer-to-message
I played around with this quite a bit, originally thinking it might be ok to expand `+select` to also accept a message, but eventually came to the conclusion that `+select` should just do what it's already doing and a new command arm should handle `;123 <message>`.

So basically the parts which relate to just rendering a message (via `;123`), I kept in `+select`, and used the functions which convert the pointer to the message for `+pointer-to-message`.

I broke out `+just-print` becuase I felt it could be used elsewhere.

Also, renamed `+activate` to `+produce` since it's now producing a noun of either the writ and index number or a reason the writ isn't there.

Finally, `+forge` was moved out for general use (mentioned in commit: https://github.com/tloncorp/landscape-apps/commit/520e675a8de0ee841d796b0f93348e0f67a2071e)