talk-cli: +reply arm to handle thread responses
This involved breaking out `+spit` (and modifying it slightly to accept a message and a `(unit id:chat`) from `+say` so that we could use it in `+reply`. The `(unit id:chat)` is added to the `memo` in our poke so that `%chat` knows whether to added it as a thread response or not.

`+reply` utilizes `+pointer-to-message` for obvious reasons, and also `+just-print`, in case we get a message indicating the message retrieval failed.