talk-cli: added $seal:chat to some arms and +render-inline adds ^
I added `seal:chat` to +build-history, +read-post, +show-post, and +mr mostly so that we can determine the first message that initiated a thread. `seal:chat` contains `replied=(set id)` so this paired with a non-null `replying=(unit id)` in the memo helps us know whether a `^` should be rendered or not.

`+render-inline` was updated to perform this check and render the `^`.