talk-cli: pointer message referencing
I added the [%reference $@(rel=@ud [zeros=@u abs=@ud]) (list inline:chat)] command and a +reference arm to handle these. It utilizes +pointer-to-message to pull the appropriate writ, packs it in as a $block, then sends it along with the reply input to %chat. The %fact we receive in return will be rendered appropriately according to the +line:mr arm, which composes the block and message reply.

I also changed two arm names for more intuitive sense:
+reply is now +thread
+spit is now +send

This also meant adding a $block type to +send so +reference can use it.

Finally, sur files for cite.hoon and groups.hoon are imported so we can cast a couple of their types in +reference.