talk-cli: threads are indented
For a cleaner distinction between them and regular messages. To do this, the +render-inline arm now will check the `$replying.memo`, and if it contains a message, it will insert a single column with nothing in it to indent the very next column.

This was more strightforward compared to my initial attempt to prepend a space within the tape itself. This was before I realized the tape essentialy text-wraps within the column so I abandoned it.