talk-cli: line wrapping & blocks for message references
I modified +block-as-tape (now +blocks-as-tapes becuase it produces a (list tape)) to render message references as dynamic, visual blocks. So now the block will adjust in relation to the width set by the user, and by the length of the sentence -- e.g. if the first and second lines within the block don't reach the max width, the longest of the two is taken as the block's max width and the shorter of the two's appended `]` is printed inline with the longest one. So visually, we get:

[ ~zod said: heya, hows it going? are ]
[ you on the other side?              ]
a reply

The implementation feels a bit convoluted, but it works.