talk-cli: glyphs displayed when targets are printed
So the user can more easily find a target's coresponding glyph. This involved updating +meta:tr to prepend the glyphs within the tapes, and using `+nome` where possible to nicely align the targets across commands which output them.

This also meant pushing the first rendered column out by two aces (should i say gap in this case? lol) so that glyphs from within messages align with the end of the line breaks.

If no glyph is assigned we simply have a gap.

You'll notice content-width is adjusted here - this was actually related to the message reference block rendering commit, as it related to widening the time column. That commit also accidentally included a portion of this commit, which was where we expanded the first column in +render-inline  from 16 to 18 to handle these extra two spaces.