talk-cli: avoid club duplicates
Some time before the message rendering portion of the appreticeship, the +read-post arm would check history and not render the post if there was a duplicate in state. Since now with each channel rendered on screen we peform a scry and add the messages to history regardless of dupes, we don't want to apply this filter. So I went ahead and incremented the `$echo` to 1 to avoid duplicate club messages. Techincally cheating, but figured this was ok for now since it's related to a bug in %chat (which has been logged).