talk-cli: using +each for +pointer-to-message
Mark tipped me to `+each`, which cleans up +pointer-to-message's output type. Also meant breaking out +index so that `+select` could use it to grab the number from absolute reference. Previously, we ouput the number as a tape in +produce and included it with the `[(unit [whom writ]) (unit tape)]`, but with `+each`, we're only producing one or the other so`+select` needs access to `+index`, independently, to get the number.