emoji picker: fix closing immediately on mobile
This stops the emoji picker from closing immediately on mobile by removing the search input.

Disabling autoFocus for the input isn't enough, if the user selects the input the picker will still close.

This stops the bleeding while we figure out what is causing the picker to close when the search within the picker is focused (or we switch to another picker or another method for displaying the picker).

Also pops the picker up closer to the message itself since we don't have to worry about the keyboard anymore.