talk-cli: refined +meta's rendering of targets
Moved +club-members to +meta and its club-id check into the %club branch within +meta since we already run a check here. We append "club doesn't exist" and "club doesn't have any members" if the check fails to produce an active club or if the `$team` and `$hive` is empty.

I dropped +render-club-title and moved the tape it produced to the bottom of the %club branch. Making a single instance of the glyph/club-id tape here enabled this.

Also, glyphs are now prepended to targets within +show-targets and +lookup-glyph to reduce +meta's complexity.