talk-cli: +poke-noun as sanity checker
Born from our discussion last week, I went ahead and implemented checks for $count = length of $history, for whether everything in view is subscribed to, and whether we have a given target in %chat. Warning messages are printed with ~&s.

I figured since it'll accept any noun, branching on type was the best method to limit what it'll do. As such, I included a sub-arm (+poke-options), which defines the commands it'll accept. This feels nice since it's self-contained.

I also moved it closer to the top of the helper core because it felt out of place below +build-history.