talk-cli: cleaned up +render-inline
Mostly to simplify and make clearer. Breaking out the part does the thread indentation did a lot; we now have as sub-arm (+thread-indent) which handles it.

I also shrunk the chat table back to 16 aces in first column so I changed a few other rendering parts to account for this.