Add presig to non-contact-store DM ships in Leap.
Currently in Leap we include ships you've DM'd with previously but aren't yet in your contact store as results.
I added this a few weeks back because we had just been including contact-store ships in the results previously.

I forgot to presig the patps :( . If you were to select one of these DM-but-not-in-contact-store ships it would navigate to a blank screen. This presigs the patps and fixes that issue.