talk-cli: +block-width refinement
Added a check to make sure the message is not null, and popped out the `t.message` check so that I could continue using `i` and `t` to refer to the tapes below.

I also got rid of `line-two`'s unit since we now check `t.message` to detemrine if the second line is empty.