talk-cli: removed ;clubs and ;targets commands
We figured since the frontend refers to clubs as group chats and the ;clubs command is the only user-facing feature that refers to them this way, we'd just get rid of it and promote clubs to ;dms. This fits well becuase the Talk frontend also lumps them together.

And, we removed ;targets because we're now left with just ;chats and ;dms to see available channels.