talk-cli: user-facing club ids are truncated
So the user can more easily identify/handle them. Since this was a UX change, I updated the +parser to accept the last 6 digits (e.g. `.loa56`) and match the full version, if it exists. If it doesn't exist, the command simply won't go through.

This also meant updating +meta:tr and +meta:mr to render truncated versions.

I kept the `.` for clarity purposes, and to match the pattern that other targets have; `~` in front of ships and `/` in front of chats.