chat: pass the care and path through a cell in +peek arms
Instead of welding the care to the end of the path, we now pass it through the +di-peek, +cu-peek, and +ca-peek arms in a cell with the path. This is so we don't have to handle the care in all cases when the path hits its final destination in dm.hoon.