talk-cli: different wires for different targets
Welp, the previous commit introduced some subscription errors. `dms` need to be on the `/ui` wire, but `clubs` and `chats` still need to be on `/ui/writs` to receive facts. I added a target type checker in +connect and +subscription-check to handle this difference. I also changed the wires in +on-agent to match these requirements to avoid confusion.