bugfix: make send button in heap detail clickable
A previous change to heap detail styles made it so that the div containing the CaretRightIcon/previous button was occluding the send button in heap detail. The div containing the left/right carets (next/prev buttons) should have had group and relative classes in order for the child divs to render/behave properly (and not occlude the items on the far left of the screen, including the send button).

Fixes #2342