bugfix: properly detect all emojis
The regex we were using before didn't account for Emoji ZWJ sequences (more detail on those here: https://emojipedia.org/emoji-zwj-sequence/), so there were some cases where a single emoji would not render in the scroller with the appropriate large font size.

I added the emoji-regex package which accounts for every type of emoji or emoji sequence. I also used lodash's `split` function to ensure that the message contains just a single rendered emoji (which the split function itself will detect as one element in the string, not multiple, even if it's actually a sequence, a feature added in lodash 4.9.0).