talk-cli: axed `;join` and `;deny` commands
Originallly added for user-facing rsvp handling, but considering this only applies to clubs and dms, we figured it's cleaner to just drop them. The agent still handles rsvp'ing, it just does so in the background when a target is put in view.

Plus, this meant we could blast the (particularly gross) +rsvp-or-print arm into oblivion.