talk-cli: use (glue ace) in %thread and %reference
Instead of `;~(pfix ace content)` because it looks better. I also moved  %thread and %reference stags to the top of +parser because if we had a target bound to either ^ or # the parser wasn't allowing us to add a space. +targ was matching the glyph before it reached these two.