talk-cli: dynamically parses club ids
I figured it would be nicer to display 12 characters for the rendered club ids - visually similar to planets - and parse these dynamically instead of having the user possibly run into a dupe. I guess this could still happen with 12 characters, but with using 5 we were going to display them in 12s anyway.

I also use the more specific +viz:ag function; however, while it will parse the full 12 characters we give it, it only passes the first 5 through the +sear gate. Wondering if this is related to the issue with +viz:ag that we found earlier...? Maybe, I'm just implementing it wrong, but I cannot discern why it's not passing the full input. Figured we could cover in our mtg.