settings: use landscape settings by default
This PR allows us to set the landscape (garden) desk settings as default by using useMergedSettings().

If a user already has a setting set in %garden, it will be honored in groups/talk until the user overrides it.

This also adds a dropdown to set the theme (auto, light, dark).