talk-cli: +message-exists uses +writ-scry-path and moved +build-history
Instead of building the path itself. The %u scry just requires dropping the `/writ` off the end of the path so we +snip after building it. Suggested here: https://github.com/tloncorp/landscape-apps/pull/2212#discussion_r1172377352

I also couldn't split this one so included the +build-history update. Since +view's the only arm that uses it, I baked it in as a sub-arm. And, I touched up the logic as it no longer uses +read-post. +read-post was used by both +build-history and +update-session, but +update-session needs to receive a (quip card session) while +build-history produces (quip card state). There's probably a way to handle this, but it felt overly complicated so I did the history/count update in the arm itself.