talk-cli: deleted +notice-remove, +show-create, and +show-delete
A fact indicating we've left a chat is passed on the `/breifs` path so we'd need to maintain a separate subscription outside of `/ui/writs` . For clubs, there are no facts that are distributed when we leave -- we'd have to check the team and hive for our ship to know. And dms have an `%archive` status, but a fact isn't distributed when it changes. Keeping %talk-cli abrest on these statuses would introduce complexity, probably not worth the benefits that +notice-remove and +show-delete would provide so i removed them.

I also got rid of +show-create becuase it wasn't being used.