Optimize chat unread initialization
This PR optimizes initialization by reducing the time taken to iterate over the briefs map and updating the state in ChatStore.

I replaced the forEach with a filter/map, which should be more performant.

I also added a new batchUnread action to the ChatStore that uses produce() update unread state in one shot, rather than calling unread on each unread brief.

This particular issue effected ships that were subscribed to a lot of channels, so it wasn't easily caught in development where we use fakezods and dev moons. On my personal ship, this saved about 15 seconds of loading time for the initial data.