talk-cli: abet pattern for session handler
The decision to convert it to an +abet pattern was made due to the complicated state and session changes we were having to perform by using +put-ses.

To build the session engine, I made the following changes:
- changed the name of +sh to +se
- passed in global state into +se so we can call outside arms that update it
- made a +split arm which handles output from arms outside of the session engine; i.e. ones that produce (quip card state). This was so we could update the global state within +se and pass the cards through +emil
- changed +make to +se-apex and updated the core-setup logic
- added an +se-abet arm
- pulled in +sh-out (now +se-out) and converted it to a stacked core within +se
- changed the necessary card-building and state-changes within each +se arm
- and adapted the arms that _call_ +se (+on-command, +command-parser, etc.) to handle passing state variables thorugh properly.

Finally, I moved +set-setting, +unset-setting and +set-timezone out of +se since they solely operate on global state.