groups: fix bug in useGroupPrivacy
Fixes #2478

We weren't properly checking the privacy of groups returned from the group index. useGroupPrivacy defaulted to `public` if we didn't have group or gang data for a particular flag, which means that we didn't show the accurate privacy setting for groups we only knew of via group index (i.e., groups we had found by searching for a ship in Find Groups).

This updates useGroupPrivacy to check the privacy status of a particular group if we can find it in the results from group index.