emojis: fix issue when picker appears below message
Fixes #2489 by removing the extra top padding around the picker.

I added the top padding in an attempt to prevent the picker from disappearing when a user accidentally moved their mouse above the picker while the picker was rendered *above* the message they are trying to react to, but didn't account for the fact that the picker sometimes renders below the message. When it appears below the message, this extra padding just pushes the picker too far and allows for a user to hover over another message and thus closes the picker.